Add scaladocs. Refactor into a single object.
